spliit uthread into main.cpp and uthread.cpp
