+++ Makefile
@@ -1,2 +1,2 @@
 uthread: main.cpp
-	g++ --std=c++11 -o uthread main.cpp -lrt
+	g++ --std=c++11 -o uthread main.cpp uthread.h uthread.cpp -lrt
